requirements.txt
websockets
websocket-client
rich
halo